interpolate/requirements.txt
autopep8==1.6.0
cycler==0.11.0
fonttools==4.34.4
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.9.0
six==1.16.0
toml==0.10.2
